feat: refactor background script wip
