+++ src/background/api/connection.ts
@@ -1,0 +1,75 @@
+import { IPermissionState } from "../../stores/reducers/permissions";
+import {
+  createAuthPopup,
+  getPermissions,
+  getStoreData,
+  setStoreData
+} from "../../utils/background";
+import { MessageFormat, validateMessage } from "../../utils/messenger";
+import { getRealURL } from "../../utils/url";
+
+/**
+ * APIs for a Web 2 like login system, but
+ * on Web 3, using the ArConnect extension
+ */
+
+// create a new connection between
+// ArConnect and the site that
+// requested it
+export const connect = (message: MessageFormat, tabURL: string) =>
+  new Promise<Partial<MessageFormat>>(async (resolve, _) => {
+    // a permission array must be submitted
+    if (!message.permissions)
+      resolve({
+        res: false,
+        message: "No permissions requested"
+      });
+
+    // check requested permissions and existing permissions
+    const existingPermissions = await getPermissions(tabURL);
+
+    // the site has a saved permission store
+    if (existingPermissions) {
+      let hasAllPermissions = true;
+
+      // if there is one permission that isn't stored in the
+      // permissions store of the url
+      // we set hasAllPermissions to false
+      for (const permission of message.permissions)
+        if (!existingPermissions.includes(permission))
+          hasAllPermissions = false;
+
+      // if all permissions are already granted we return
+      // TODO: maybe return permissions here?
+      if (hasAllPermissions)
+        resolve({
+          res: false,
+          message: "All permissions are already allowed for this site"
+        });
+    }
+
+    createAuthPopup({
+      permissions: message.permissions,
+      type: "connect",
+      url: tabURL
+    });
+    chrome.runtime.onMessage.addListener((msg) => {
+      if (validateMessage(msg, { sender: "popup", type: "connect_result" }))
+        resolve(msg);
+    });
+  });
+
+export const disconnect = (tabURL: string) =>
+  new Promise<Partial<MessageFormat>>(async (resolve, _) => {
+    try {
+      const store = await getStoreData();
+      await setStoreData({
+        permissions: (store.permissions ?? []).filter(
+          (sitePerms: IPermissionState) => sitePerms.url !== getRealURL(tabURL)
+        )
+      });
+      resolve({ res: true });
+    } catch {
+      resolve({ res: false, message: "Could not disconnect" });
+    }
+  });
